wai-aria/tree/basic.tentative.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/wai-aria/scripts/aria-utils.js"></script>

<div id="listbox" role="listbox" aria-label="listbox">
  <div id="option1" role="option" aria-label="option1"></div>
  <div id="option2" role="option" aria-label="option2"></div>
</div>

<script>

promise_test(async t => {
  const listbox = await test_driver.get_element_accessible_properties(document.getElementById("listbox"));
  assert_equals(listbox.children.length, 2);
  const option1 = await test_driver.get_accessible_properties(listbox.children[0]);
  assert_equals(option1.id, listbox.children[0]);
  assert_equals(option1.parent, listbox.id);
  assert_equals(option1.role, "option");
  assert_equals(option1.label, "option1");
  const option2 = await test_driver.get_accessible_properties(listbox.children[1]);
  assert_equals(option2.id, listbox.children[1]);
  assert_equals(option2.parent, listbox.id);
  assert_equals(option2.label, "option2");
}, "listbox explicit tree");

AriaUtils.verifyAccessibilityTree("#listbox", {
  role: "listbox",
  label: "listbox",
  children: [
    { role: "option", label: "option1", children: [] },
    { role: "option", label: "option2", children: [] },
  ],
});

</script>
